+++ src/doc/resources/bridlensis/doc/release_notes.md
@@ -6,6 +6,7 @@ Release date: TBD
 
 ### What's New
 
+*   New convenience functions for Push, Pop, Exch, StrCmp and IntCmp
 *   New convenience functions for file function headers as defined in NSIS 3.0a2
 *   Added Apache Ant task to compile BridleNSIS script installers